# api/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for song management",
        "title": "Song API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "David Slatinek",
            "url": "https://github.com/david-slatinek"
        },
        "license": {
            "name": "Apache-2.0 license",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/songs": {
            "get": {
                "description": "Get all songs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Get all songs",
                "responses": {
                    "200": {
                        "description": "An array of songs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SongDto"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new song.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Create a new song",
                "parameters": [
                    {
                        "description": "Song info",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/SongDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "description": "Get a song by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Get a song by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SongDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update song.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "song"
                ],
                "summary": "Update song",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated song",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SongDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "description": "Error model",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "Song": {
            "description": "Model for creating a new song",
            "type": "object",
            "required": [
                "author",
                "duration",
                "release",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "release": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "SongDto": {
            "description": "Model for a created song",
            "type": "object",
            "required": [
                "author",
                "duration",
                "release",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "release": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}